Fix bundled bitmaps not being unloaded
When loading, the IMAGE_RESOURCE case had the following check:
if ( ! has_resource_bundle(line_name) ) return;
This would always return (because it was checking resource bundles
rather than bitmaps), bitmaps to be loaded but never be added to
the list of bundled resources.
This in turn resulted in free_resource_bundle never freeing that resource.